+++ Source/DataSources/PolygonGeometryUpdater.js
@@ -8,8 +8,8 @@ define([
         '../Core/destroyObject',
         '../Core/DeveloperError',
         '../Core/Event',
-        '../Core/isArray',
         '../Core/GeometryInstance',
+        '../Core/isArray',
         '../Core/Iso8601',
         '../Core/PolygonGeometry',
         '../Core/PolygonHierarchy',
@@ -31,8 +31,8 @@ define([
         destroyObject,
         DeveloperError,
         Event,
-        isArray,
         GeometryInstance,
+        isArray,
         Iso8601,
         PolygonGeometry,
         PolygonHierarchy,
@@ -518,7 +518,6 @@ define([
         this._outlinePrimitive = undefined;
         this._geometryUpdater = geometryUpdater;
         this._options = new GeometryOptions(geometryUpdater._entity);
-        this._scrach = undefined;
     };
 
     DynamicGeometryUpdater.prototype.update = function(time) {
@@ -540,7 +539,7 @@ define([
         }
 
         var options = this._options;
-        var hierarchy = Property.getValueOrUndefined(polygon.hierarchy, time, this._scrach);
+        var hierarchy = Property.getValueOrUndefined(polygon.hierarchy, time);
         if (!defined(hierarchy)) {
             return;
         }